Renamed GMP base operator module to match product naming conventions.
